# [Numpy] FFI for einsum, dstack, unique