@haojin2 2020-03-25
@mxnet-bot run ci [unix-cpu, unix-gpu]